lab6_telegram_bot/requirements.txt
aiogram==2.23.1
Pillow==9.3.0 